Moved node_util.py module from input package to util package
