+++ src/pds_doi_service/core/util/doi_xml_differ.py
@@ -10,8 +10,8 @@ import datetime
 
 from lxml import etree
 from pds_doi_service.core.entities.exceptions import InputFormatException
-from pds_doi_service.core.input.node_util import NodeUtil
 from pds_doi_service.core.util.general_util import get_logger
+from pds_doi_service.core.util.node_util import NodeUtil
 
 logger = get_logger("pds_doi_core.util.doi_xml_differ")
