+++ src/pds_doi_service/core/util/node_util.py
